+++ Store/admin/components/Product/Edit.php
@@ -3,7 +3,6 @@
 require_once 'Admin/pages/AdminDBEdit.php';
 require_once 'Admin/exceptions/AdminNotFoundException.php';
 require_once 'Admin/exceptions/AdminNoAccessException.php';
-require_once 'NateGoSearch/NateGoSearch.php';
 require_once 'Swat/SwatDate.php';
 require_once 'SwatDB/SwatDB.php';
 require_once 'Store/dataobjects/StoreProduct.php';
@@ -15,7 +14,7 @@ require_once 'Store/StoreCatalogSelector.php';
  * Edit page for Products
  *
  * @package   Store
- * @copyright 2005-2007 silverorange
+ * @copyright 2005-2012 silverorange
  * @license   http://www.gnu.org/copyleft/lesser.html LGPL License 2.1
  */
 class StoreProductEdit extends AdminDBEdit
@@ -253,10 +252,17 @@ class StoreProductEdit extends AdminDBEdit
 
 	protected function addToSearchQueue()
 	{
+		$manager = $this->app->db->manager;
+		if (!in_array('nategosearchqueue', $manager->listTables())) {
+			return;
+		}
+
+		require_once 'NateGoSearch/NateGoSearch.php';
 		$type = NateGoSearch::getDocumentType($this->app->db, 'product');
 
-		if ($type === null)
+		if ($type === null) {
 			return;
+		}
 
 		$sql = sprintf('delete from NateGoSearchQueue
 			where document_id = %s and document_type = %s',
